UB fixes provided by rice123
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@46198 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/snapshot.c
@@ -271,7 +271,7 @@ static int snapshot_read_dword(FILE *f, uint32_t *dw_return)
         return -1;
     }
 
-    *dw_return = lo | (hi << 16);
+    *dw_return = lo | ((uint32_t)hi << 16);
     return 0;
 }
